# routers/api/v1/tag.go
package v1

import (
  "log"
  "net/http"
  "github.com/gin-gonic/gin"
  "github.com/beego/beego/v2/core/validation"

  "github.com/unknwon/com"
  "gin_tutorial/models"
  "gin_tutorial/pkg/e"
  "gin_tutorial/pkg/setting"
  "gin_tutorial/pkg/util"
)

// @Summary 複数記事タグ取得
// @Produce  json
// @Param name query string false "Name"
// @Param state query int false "State"
// @Success 200 {string} json "{"code":200,"data":{"lists":[{"id":3,"created_on":1516849721,"modified_on":0,"name":"3333","created_by":"4555","modified_by":"","state":0}],"total":29},"msg":"ok"}"
// @Router /api/v1/tags [get]
func GetTags(c *gin.Context) {
  name := c.Query("name")
  maps := make(map[string]interface{})
  data := make(map[string]interface{})

  if name != "" {
    maps["name"] = name
  }

  var state int = -1
  if arg := c.Query("state"); arg != "" {
    state = com.StrTo(arg).MustInt()
    maps["state"] = state
  }
  code := e.SUCCESS
  data["lists"] = models.GetTags(util.GetPage(c), setting.Config.PageSize, maps)
  data["total"] = models.GetTagTotal(maps)
  c.JSON(http.StatusOK, gin.H{
    "code": code,
    "msg": e.GetMsg(code),
    "data": data,
  })
}

// @Summary 新規記事タク
// @Produce  json
// @Param name query string true "Name"
// @Param state query int false "State"
// @Param created_by query int false "CreatedBy"
// @Success 200 {string} json "{"code":200,"data":{},"msg":"ok"}"
// @Router /api/v1/tags [post]
func AddTag(c *gin.Context) {
  name      := c.Query("name")
  state     := com.StrTo(c.DefaultQuery("state", "0")).MustInt()
  createdBy := c.Query("created_by")
  valid := validation.Validation{}
  valid.Required(name, "name").Message("The name cannot be empty")
  valid.MaxSize(name, 100, "name").Message("The name is a maximum of 100 characters")
  valid.Required(createdBy, "created_by").Message("The creator cannot be empty")
  valid.MaxSize(createdBy, 100, "created_by").Message("The creator is up to 100 characters long")
  valid.Range(state, 0, 1, "state").Message("Status only allows 0 or 1")
  code := e.INVALID_PARAMS
  if ! valid.HasErrors() {
    if ! models.ExistTagByName(name) {
      code = e.SUCCESS
      models.AddTag(name, state, createdBy)
    } else {
      code = e.ERROR_EXIST_TAG
    }
  } else {
    for _, err := range valid.Errors {
      log.Println(err.Key, err.Message)
    }
  }
  c.JSON(http.StatusOK, gin.H{
    "code" : code,
    "msg"  : e.GetMsg(code),
    "data" : make(map[string]string),
  })
}

// @Summary 修正記事タク
// @Produce  json
// @Param id path int true "ID"
// @Param name query string true "ID"
// @Param state query int false "State"
// @Param modified_by query string true "ModifiedBy"
// @Success 200 {string} json "{"code":200,"data":{},"msg":"ok"}"
// @Router /api/v1/tags/{id} [put]
func EditTag(c *gin.Context) {
  id := com.StrTo(c.Param("id")).MustInt()
  name := c.Query("name")
  modifiedBy := c.Query("modified_by")

  valid := validation.Validation{}

  var state int = -1
  if arg := c.Query("name"); arg != "" {
      state = com.StrTo(arg).MustInt()
      valid.Range(state, 0, 1, "state").Message("Status only allows 0 or 1")
  }
  valid.Required(id, "id").Message("ID cannot be empty")
  valid.Required(modifiedBy, "modified_by").Message("The creator cannot be empty")
  valid.MaxSize(modifiedBy, 100, "modified_by").Message("The creator is up to 100 characters long")
  valid.MaxSize(name, 100, "name").Message("The name is a maximum of 100 characters")

  code := e.INVALID_PARAMS
  if ! valid.HasErrors() {
      code = e.SUCCESS
      if models.ExistTagByID(id) {
          data := make(map[string]interface{})
          data["modified_by"]  = modifiedBy
          if name != "" {
              data["name"] = name
          }
          models.EditTag(id, data)
      } else {
          code = e.ERROR_NOT_EXIST_TAG
      }
  } else {
      for _, err := range valid.Errors {
          log.Println(err.Key, err.Message)
      }
  }
  c.JSON(http.StatusOK, gin.H{
    "code" : code,
    "msg"  : e.GetMsg(code),
    "data" : make(map[string]string),
  })
}

// @Summary 删除記事タク
// @Produce  json
// @Param id path int true "ID"
// @Success 200 {string} json "{"code":200,"data":{},"msg":"ok"}"
// @Router /api/v1/tags/{id} [delete]
func DeleteTag(c *gin.Context) {
  id := com.StrTo(c.Param("id")).MustInt()

  valid := validation.Validation{}
  valid.Min(id, 1, "id").Message("ID must be greater than 0")

  code := e.INVALID_PARAMS
  if ! valid.HasErrors() {
    code = e.SUCCESS
    if models.ExistTagByID(id) {
      models.DeleteTag(id)
    } else {
      code = e.ERROR_NOT_EXIST_TAG
    }
  } else {
    for _, err := range valid.Errors {
      log.Println(err.Key, err.Message)
    }
  }
  
  c.JSON(http.StatusOK, gin.H{
    "code" : code,
    "msg"  : e.GetMsg(code),
    "data" : make(map[string]string),
  })
}